+++ core/src/main/kotlin/dev/zieger/utils/coroutines/channel/PipeValueFiFo.kt
@@ -9,6 +9,6 @@ import dev.zieger.utils.misc.lastOrNull
 
 open class PipeValueFiFo<T>(amount: Int) : FiFo<IPipeValue<T>>(amount) {
 
-    open fun putValue(value: IPipeValue<T>) = put(value, value.time == values.lastOrNull()?.time)
+    open fun putValue(value: IPipeValue<T>) = put(value, value.time == lastOrNull()?.time)
 }